% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_abfrage_fundamentaldaten.R
\name{abfrage_fundamentaldaten}
\alias{abfrage_fundamentaldaten}
\title{abfrage_fundamentaldaten}
\usage{
abfrage_fundamentaldaten(url)
}
\arguments{
\item{url}{character - URL zu Unternehmens-Daten}
}
\value{
data.table enthält alle Fundamentaldaten der letzten 12 Jahre
}
\description{
Die Funktion ruft die Fundamentaldaten eines Unternehmens ab
}
